AGPL3.0 → PostgreSQL licence change
I feel that the PostgreSQL licence, after all, is a more suitable
licence to be a team player within the PostgreSQL ecosystem (regardless
of my more enduring love of the FSF their beautiful viral licence).  And
now that I'm still the only author, a licence change is a trivial
affair.